added iOS modules build

+++ modules/driver.sh
@@ -7,6 +7,13 @@ export PATH=/home/hudson/linux_slave/bin:/opt/apache-ant/bin:$PATH
 export PKG_CONFIG_PATH=/home/hudson/linux_slave/lib/pkgconfig:$PKG_CONFIG_PATH
 export LD_LIBRARY_PATH=/home/hudson/linux_slave/lib
 
+#export TITANIUM_BUILD=/home/hudson/Source/titanium_build
+export TITANIUM_BUILD=/Users/nikolai/build/titanium_build
+#export WORKSPACE=/home/hudson/linux_slave/stage/titanium_mobile_modules
+export WORKSPACE=/Users/nikolai/build/titanium_mobile_modules
+
+cd $WORKSPACE
+
 GIT_BRANCH=$1
 GIT_REVISION=`git log --pretty=oneline -n 1 | sed 's/ .*//' | tr -d '\n' | tr -d '\r'`
 #VERSION=`python $TITANIUM_BUILD/common/get_version.py | tr -d '\r'`
@@ -17,31 +24,46 @@ if [ "$PYTHON" = "" ]; then
 	PYTHON=python
 fi
 
-# Android modules
+# iOS modules
 
-#export TITANIUM_BUILD=/home/hudson/Source/titanium_build
-#export WORKSPACE=/home/hudson/linux_slave/stage/titanium_mobile_modules
+for MODULE in `ls $WORKSPACE/iphone`; do
+	cd $WORKSPACE/iphone/$MODULE
+	echo "building $MODULE (iOS)..."
+	./build.py  >& build.log
+	grep 'BUILD SUCCEEDED'  build.log | uniq
+	grep -q 'BUILD SUCCEEDED' build.log || cat build.log
+	# check if zip was updated
+	ZIP=`ls *\.zip`
+	if [ -z "$ZIP" ]; then continue; fi
+	STAMPED_ZIP=`echo $ZIP| sed "s/\/\(.*\).zip/\/\1-$TIMESTAMP.zip/"`
+	mv $ZIP $STAMPED_ZIP
+	#$PYTHON $TITANIUM_BUILD/common/s3_uploader.py modules $STAMPED_ZIP $GIT_BRANCH $GIT_REVISION $BUILD_URL
+	rm $STAMPED_ZIP
+done
+
+exit
+# Android modules
 
 for MODULE in `ls $WORKSPACE/android`; do
 	MIN_SDK=`grep minsdk $WORKSPACE/android/$MODULE/manifest | cut -f2 -d' '`
 	sed s/TITANIUM_VERSION/$MIN_SDK/ $TITANIUM_BUILD/modules/build.properties.template > $WORKSPACE/android/$MODULE/build.properties
 	cd $WORKSPACE/android/$MODULE
-	echo building $MODULE...
-	MOD_TIME_PREV=`ls -l --time-style=+'%Y%m%d%H%M%S' dist/*\.zip | awk '{print $6}'`
+	echo "building $MODULE (Android)..."
+	#MOD_TIME_PREV=`ls -l --time-style=+'%Y%m%d%H%M%S' dist/*\.zip | awk '{print $6}'`
 	ZIP=`ls dist/*\.zip`
-	rm $ZIP
-	ant 2>&1 > ant.log
-	MOD_TIME_POST=`ls -l --time-style=+'%Y%m%d%H%M%S' dist/*\.zip | awk '{print $6}'`
+	if [ "$ZIP" ]; then rm $ZIP; fi
+	ant #2>&1 > ant.log
+	#MOD_TIME_POST=`ls -l --time-style=+'%Y%m%d%H%M%S' dist/*\.zip | awk '{print $6}'`
 	# check if zip was updated
 	#echo $MOD_TIME_PREV
 	#echo $MOD_TIME_POST
 	#if [ "$MOD_TIME_PREV" = "$MOD_TIME_POST" ]; then continue; fi
 	ZIP=`ls dist/*\.zip`
 	if [ -z "$ZIP" ]; then continue; fi
-	STAMPED_ZIP=`echo $ZIP| sed "s/\/\(.*\).zip/\/\1-$MOD_TIME_POST.zip/"`
-	cp $ZIP $STAMPED_ZIP
-	$PYTHON $TITANIUM_BUILD/common/s3_uploader.py modules $STAMPED_ZIP $GIT_BRANCH $GIT_REVISION $BUILD_URL
-	rm $STAMPED_ZIP
+	STAMPED_ZIP=`echo $ZIP| sed "s/\/\(.*\).zip/\/\1-$TIMESTAMP.zip/"`
+	mv $ZIP $STAMPED_ZIP
+	#$PYTHON $TITANIUM_BUILD/common/s3_uploader.py modules $STAMPED_ZIP $GIT_BRANCH $GIT_REVISION $BUILD_URL
+	#rm $STAMPED_ZIP
 done
 
 #$PYTHON $TITANIUM_BUILD/common/s3_cleaner.py modules $GIT_BRANCH
